refac npc & player shop opening
