+++ src/core/interface/networking/packets/packet/in/onclick/OnClickPacketHandler.ts
@@ -4,37 +4,26 @@ import Logger from '@/core/infra/logger/Logger';
 import OnClickPacket from './OnClickPacket';
 import { QuestManager } from '@/core/domain/quests/QuestManager';
 import NPC from '@/core/domain/entities/game/mob/NPC';
-import Player from '@/core/domain/entities/game/player/Player';
-import ShopService from '@/game/app/service/ShopService';
 import ShopManager from '@/core/domain/shop/ShopManager';
-import PrivateShopService from '@/game/app/service/PrivateShopService';
 
 export default class OnClickPacketHandler extends PacketHandler<OnClickPacket> {
     private readonly logger: Logger;
     private readonly questManager: QuestManager;
-    private readonly shopService: ShopService;
     private readonly shopManager: ShopManager;
-    private readonly privateShopService: PrivateShopService;
 
     constructor({
         logger,
         questManager,
         shopManager,
-        shopService,
-        privateShopService,
     }: {
         logger: Logger;
         questManager: QuestManager;
         shopManager: ShopManager;
-        shopService: ShopService;
-        privateShopService: PrivateShopService;
     }) {
         super();
         this.logger = logger;
         this.questManager = questManager;
-        this.shopService = shopService;
         this.shopManager = shopManager;
-        this.privateShopService = privateShopService;
     }
 
     async execute(connection: GameConnection, packet: OnClickPacket) {
@@ -73,30 +62,12 @@ export default class OnClickPacketHandler extends PacketHandler<OnClickPacket> {
             return;
         }
 
-        // If the clicked entity is a player running a private shop, open it for the guest
-        if (target instanceof Player) {
-            if (target.isRunningPrivateShop()) {
-                if (target === player) {
-                    // Owner clicked themselves - re-send the shop listing to open the management UI
-                    await this.privateShopService.openShopForOwner(player);
-                    return;
-                }
-                this.logger.debug(
-                    `[OnClickPacketHandler] ${player.getName()} clicked private shop owner: ${target.getName()}`,
-                );
-                await this.privateShopService.openShopForGuest(player, target);
-            }
-            return;
-        }
-
         const npc = target as NPC;
         this.logger.info(`[OnClickPacketHandler] You have clicked on: ${npc.getId()}`);
 
         // If the npc has a quest and is a shop, then do the quest first and skip opening the shop
         this.questManager.onClick(player, npc);
 
-        if (this.shopManager.hasShop(npc.getId())) {
-            await this.shopService.openShop(player, npc);
-        }
+        await this.shopManager.openShop(target, player);
     }
 }
