+++ src/game/app/GameApplication.ts
@@ -9,7 +9,7 @@ import CacheProvider from '@/core/infra/cache/CacheProvider';
 import DatabaseManager from '@/core/infra/database/DatabaseManager';
 import Server from '@/core/interface/server/Server';
 import Logger from '@/core/infra/logger/Logger';
-import ShopManager from '@/core/domain/shop/ShopManager';
+import ShopService from '@/game/app/service/ShopService';
 
 export default class GameApplication extends Application {
     private readonly world: World;
@@ -18,7 +18,7 @@ export default class GameApplication extends Application {
     private readonly itemManager: ItemManager;
     private readonly dropManager: DropManager;
     private readonly questManager: QuestManager;
-    private readonly shopManager: ShopManager;
+    private readonly shopService: ShopService;
 
     constructor(container: {
         logger: Logger;
@@ -31,7 +31,7 @@ export default class GameApplication extends Application {
         itemManager: ItemManager;
         dropManager: DropManager;
         questManager: QuestManager;
-        shopManager: ShopManager;
+        shopService: ShopService;
     }) {
         super(container);
         this.world = container.world;
@@ -40,7 +40,7 @@ export default class GameApplication extends Application {
         this.itemManager = container.itemManager;
         this.dropManager = container.dropManager;
         this.questManager = container.questManager;
-        this.shopManager = container.shopManager;
+        this.shopService = container.shopService;
     }
 
     async start() {
@@ -54,7 +54,7 @@ export default class GameApplication extends Application {
         ]);
         this.mobManager.load();
         this.questManager.load();
-        this.shopManager.load();
+        this.shopService.load();
         await this.server.setup().start();
         await this.world.init();
         this.logger.info('[APP] Game application started 🎮🚀');
