+++ src/game/app/service/ShopService.ts
@@ -1,158 +1,66 @@
-import Player from '@/core/domain/entities/game/player/Player';
-import NPC from '@/core/domain/entities/game/mob/NPC';
-import ShopManager from '@/core/domain/shop/ShopManager';
-import ItemManager from '@/core/domain/manager/ItemManager';
+import Shop from '@/core/domain/shop/Shop';
+import { ShopItem } from '@/core/domain/shop/ShopItem';
 import Logger from '@/core/infra/logger/Logger';
-import { PointsEnum } from '@/core/enum/PointsEnum';
-import { ItemAntiFlagEnum } from '@/core/enum/ItemAntiFlagEnum';
-import { ShopSubHeaderGC } from '@/core/enum/ShopSubHeaderEnum';
-import { WindowTypeEnum } from '@/core/enum/WindowTypeEnum';
-import PrivateShopService from './PrivateShopService';
+import ItemManager from '@/core/domain/manager/ItemManager';
+import { GameConfig } from '@/game/infra/config/GameConfig';
+
+type ShopServiceParams = {
+    config: GameConfig;
+    logger: Logger;
+    itemManager: ItemManager;
+};
 
 export default class ShopService {
-    private readonly shopManager: ShopManager;
-    private readonly itemManager: ItemManager;
-    private readonly privateShopService: PrivateShopService;
+    private readonly config: GameConfig;
     private readonly logger: Logger;
+    private readonly itemManager: ItemManager;
+    private readonly shops: Map<number, Shop> = new Map();
 
-    constructor({
-        shopManager,
-        itemManager,
-        privateShopService,
-        logger,
-    }: {
-        shopManager: ShopManager;
-        itemManager: ItemManager;
-        privateShopService: PrivateShopService;
-        logger: Logger;
-    }) {
-        this.shopManager = shopManager;
-        this.itemManager = itemManager;
-        this.privateShopService = privateShopService;
+    constructor({ config, logger, itemManager }: ShopServiceParams) {
+        this.config = config;
         this.logger = logger;
+        this.itemManager = itemManager;
     }
 
-    async openShop(player: Player, npc: NPC, id?: number) {
-        const shop = this.shopManager.getShop(id ?? npc.getId());
-        if (!shop) {
-            this.logger.debug(`[ShopService] NPC vnum ${npc.getId()} has no shop`);
-            return;
-        }
+    load() {
+        this.config.npcShops.forEach((shopEntry) => {
+            const shopItems: ShopItem[] = [];
 
-        //TODO: validate if player can open the shop (distance, isExchanging, HasShopOpened, etc)
-        //TODO: verify player distance to npc
+            shopEntry.items.forEach(({ vnum, count, position }, index) => {
+                const item = this.itemManager.getItem(vnum, count);
+                if (!item) {
+                    this.logger.info(`[ShopManager] Unknown item vnum ${vnum} in shop for NPC ${shopEntry.npcVnum}`);
+                    return;
+                }
 
-        player.setCurrentShop(shop);
+                shopItems.push({
+                    vnum,
+                    count,
+                    price: item.getShopPrice(),
+                    item,
+                    position: position ?? index,
+                    size: item.getSize(),
+                });
+            });
 
-        player.sendCurrentShop({
-            ownerVid: npc.getVirtualId(),
-            items: shop.getItems(),
+            const shop = new Shop({
+                npcVnum: shopEntry.npcVnum,
+                shopName: shopEntry.shopName,
+                items: shopItems,
+            });
+
+            this.shops.set(shopEntry.npcVnum, shop);
+            this.logger.info(
+                `[ShopManager] Loaded shop for NPC vnum ${shopEntry.npcVnum} "${shopEntry.shopName}" with ${shopItems.length} item(s)`,
+            );
         });
-
-        this.logger.info(
-            `[ShopService] Player ${player.getName()} opened shop "${shop.getShopName()}" (NPC ${npc.getId()})`,
-        );
     }
 
-    async closeShop(player: Player) {
-        // If browsing a private shop, leave it gracefully
-        const privateOwner = player.getCurrentPrivateShopOwner();
-        if (privateOwner) {
-            privateOwner.getPrivateShop()?.removeGuest(player);
-            player.setCurrentPrivateShopOwner(null);
-            player.sendShopClose();
-            this.logger.info(`[ShopService] Player ${player.getName()} left private shop of ${privateOwner.getName()}`);
-            return;
-        }
-
-        if (!player.getCurrentShop()) return;
-
-        player.setCurrentShop(null);
-        player.sendShopClose();
-
-        this.logger.info(`[ShopService] Player ${player.getName()} closed shop`);
+    hasShop(npcVnum: number): boolean {
+        return this.shops.has(npcVnum);
     }
 
-    async buy(player: Player, pos: number) {
-        // Route to private shop service if the player is browsing a private shop
-        if (player.getCurrentPrivateShopOwner()) {
-            await this.privateShopService.buy(player, pos);
-            return;
-        }
-
-        const shop = player.getCurrentShop();
-        if (!shop) {
-            this.logger.debug(`[ShopService] buy: player ${player.getName()} has no open shop`);
-            return;
-        }
-
-        const shopItem = shop.getItem(pos);
-        if (!shopItem) {
-            player.sendShopResult({ result: ShopSubHeaderGC.INVALID_POS });
-            return;
-        }
-
-        const price = shopItem.price;
-        const playerGold = player.getPoint(PointsEnum.GOLD);
-
-        if (playerGold < price) {
-            player.sendShopResult({ result: ShopSubHeaderGC.NOT_ENOUGH_MONEY });
-            return;
-        }
-
-        // Build a fresh item instance from proto
-        const item = this.itemManager.getItem(shopItem.vnum, shopItem.count);
-        if (!item) {
-            player.sendShopResult({ result: ShopSubHeaderGC.SOLD_OUT });
-            return;
-        }
-
-        if (!player.addItem(item)) {
-            player.sendShopResult({ result: ShopSubHeaderGC.INVENTORY_FULL });
-            return;
-        }
-
-        player.addPoint(PointsEnum.GOLD, -price);
-        await this.itemManager.save(item);
-
-        player.sendShopResult({ result: ShopSubHeaderGC.OK });
-
-        this.logger.info(
-            `[ShopService] Player ${player.getName()} bought vnum ${shopItem.vnum} x${shopItem.count} for ${price}g`,
-        );
-    }
-
-    async sell(player: Player, pos: number, count: number) {
-        const shop = player.getCurrentShop();
-        if (!shop) {
-            this.logger.debug(`[ShopService] sell: player ${player.getName()} has no open shop`);
-            return;
-        }
-
-        const item = player.getItem(pos);
-        if (!item) {
-            player.sendShopResult({ result: ShopSubHeaderGC.INVALID_POS });
-            return;
-        }
-
-        // Cannot sell items flagged as anti-sell
-        if (item.getAntiFlags().is(ItemAntiFlagEnum.ANTI_SELL)) {
-            player.sendShopResult({ result: ShopSubHeaderGC.INVALID_POS });
-            return;
-        }
-
-        const sellCount = Math.min(count, item.getCount() ?? 1);
-        const sellPrice = Math.floor((item.getShopPrice() * sellCount) / 5);
-
-        // Remove item from inventory and notify client
-        player.getInventory().removeItem(pos, item.getSize());
-        player.sendItemRemoved({ window: WindowTypeEnum.INVENTORY, position: pos });
-        await this.itemManager.delete(item);
-
-        player.addPoint(PointsEnum.GOLD, sellPrice);
-
-        player.sendShopResult({ result: ShopSubHeaderGC.OK });
-
-        this.logger.info(`[ShopService] Player ${player.getName()} sold pos=${pos} x${sellCount} for ${sellPrice}g`);
+    getShop(npcVnum: number): Shop | undefined {
+        return this.shops.get(npcVnum);
     }
 }
