+++ test/unit/core/domain/manager/ShopManager.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 import WinstonLoggerAdapter from '@/core/infra/logger/WinstonLoggerAdapter';
-import ShopManager from '@/core/domain/shop/ShopManager';
+import ShopService from '@/game/app/service/ShopService';
 import Shop from '@/core/domain/shop/Shop';
 import ItemManager from '@/core/domain/manager/ItemManager';
 import { makeGameConfig } from '@/game/infra/config/GameConfig';
@@ -9,13 +9,13 @@ import { makeGameConfig } from '@/game/infra/config/GameConfig';
 describe('ShopManager', () => {
     let loggerStub: sinon.SinonStubbedInstance<WinstonLoggerAdapter>;
     let itemManagerStub: sinon.SinonStubbedInstance<ItemManager>;
-    let manager: ShopManager;
+    let service: ShopService;
 
     beforeEach(() => {
         loggerStub = sinon.createStubInstance(WinstonLoggerAdapter);
         itemManagerStub = sinon.createStubInstance(ItemManager);
         itemManagerStub.getItem.returns({ getShopPrice: () => 100, getSize: () => 1 } as any);
-        manager = new ShopManager({
+        service = new ShopService({
             config: makeGameConfig(),
             logger: loggerStub,
             itemManager: itemManagerStub,
@@ -27,27 +27,27 @@ describe('ShopManager', () => {
     });
 
     it('should initialize with empty shops', () => {
-        expect(manager.hasShop(9001)).to.be.false;
+        expect(service.hasShop(9001)).to.be.false;
     });
 
     it('should load shops from npc_shop.json', () => {
         // Mock itemManager.getItem to return valid items
         itemManagerStub.getItem.returns({ getShopPrice: () => 100, getSize: () => 1 } as any);
 
-        manager.load();
+        service.load();
 
         // Verify shops are loaded
-        expect(manager.hasShop(9001)).to.be.true;
-        expect(manager.hasShop(9002)).to.be.true;
-        expect(manager.hasShop(9012)).to.be.true;
+        expect(service.hasShop(9001)).to.be.true;
+        expect(service.hasShop(9002)).to.be.true;
+        expect(service.hasShop(9012)).to.be.true;
     });
 
     it('should return correct shop for valid NPC vnum', () => {
         itemManagerStub.getItem.returns({ getShopPrice: () => 100, getSize: () => 1 } as any);
 
-        manager.load();
+        service.load();
 
-        const shop = manager.getShop(9001);
+        const shop = service.getShop(9001);
         expect(shop).to.be.instanceOf(Shop);
         if (shop) {
             expect(shop.getShopName()).to.equal('Arms');
@@ -57,38 +57,38 @@ describe('ShopManager', () => {
     it('should return undefined for invalid NPC vnum', () => {
         itemManagerStub.getItem.returns({ getShopPrice: () => 100, getSize: () => 1 } as any);
 
-        manager.load();
+        service.load();
 
-        const shop = manager.getShop(9999);
+        const shop = service.getShop(9999);
         expect(shop).to.be.undefined;
     });
 
     it('should indicate shop exists for loaded NPCs', () => {
         itemManagerStub.getItem.returns({ getShopPrice: () => 100, getSize: () => 1 } as any);
 
-        manager.load();
+        service.load();
 
-        expect(manager.hasShop(9001)).to.be.true;
-        expect(manager.hasShop(9999)).to.be.false;
+        expect(service.hasShop(9001)).to.be.true;
+        expect(service.hasShop(9999)).to.be.false;
     });
 
     it('should load all expected shops', () => {
         itemManagerStub.getItem.returns({ getShopPrice: () => 100, getSize: () => 1 } as any);
 
-        manager.load();
+        service.load();
 
         const expectedNpcs = makeGameConfig().npcShops.map((shopEntry) => shopEntry.npcVnum);
         expectedNpcs.forEach((npcVnum) => {
-            expect(manager.hasShop(npcVnum)).to.be.true;
+            expect(service.hasShop(npcVnum)).to.be.true;
         });
     });
 
     it('should load shop items correctly', () => {
         itemManagerStub.getItem.returns({ getShopPrice: () => 500, getSize: () => 1 } as any);
 
-        manager.load();
+        service.load();
 
-        const shop = manager.getShop(9001);
+        const shop = service.getShop(9001);
         expect(shop).to.not.be.undefined;
         if (shop) {
             const items = shop.getItems();
