+++ test/unit/game/app/service/ShopService.test.ts
@@ -1,7 +1,7 @@
 import { expect } from 'chai';
 import sinon from 'sinon';
 import WinstonLoggerAdapter from '@/core/infra/logger/WinstonLoggerAdapter';
-import ShopService from '@/game/app/service/ShopService';
+import ShopManager from '@/core/domain/shop/ShopManager';
 import Shop from '@/core/domain/shop/Shop';
 import { ShopItem } from '@/core/domain/shop/ShopItem';
 import { ShopSubHeaderGC } from '@/core/enum/ShopSubHeaderEnum';
@@ -11,7 +11,7 @@ describe('ShopService', () => {
     let loggerStub;
     let shopManagerStub;
     let itemManagerStub;
-    let service: ShopService;
+    let manager: ShopManager;
     let playerStub;
     let npcStub;
 
@@ -20,9 +20,9 @@ describe('ShopService', () => {
         shopManagerStub = { getShop: sinon.stub(), hasShop: sinon.stub() };
         itemManagerStub = { getItem: sinon.stub(), save: sinon.stub().resolves(), delete: sinon.stub().resolves() };
 
-        service = new ShopService({
+        manager = new ShopManager({
             logger: loggerStub,
-            shopManager: shopManagerStub,
+            shopService: shopManagerStub,
             itemManager: itemManagerStub,
             privateShopService: {} as any,
         });
@@ -61,7 +61,7 @@ describe('ShopService', () => {
             const shop = new Shop({ npcVnum: 9001, shopName: 'Arms', items });
             shopManagerStub.getShop.returns(shop);
 
-            await service.openShop(playerStub, npcStub);
+            await manager.openShop(playerStub, npcStub);
 
             expect(playerStub.setCurrentShop.calledOnceWith(shop)).to.be.true;
             expect(playerStub.sendCurrentShop.calledOnce).to.be.true;
@@ -70,7 +70,7 @@ describe('ShopService', () => {
         it('should not open shop if NPC has no shop', async () => {
             shopManagerStub.getShop.returns(undefined);
 
-            await service.openShop(playerStub, npcStub);
+            await manager.openShop(playerStub, npcStub);
 
             expect(playerStub.setCurrentShop.called).to.be.false;
             expect(playerStub.sendCurrentShop.called).to.be.false;
@@ -82,7 +82,7 @@ describe('ShopService', () => {
             const shop = new Shop({ npcVnum: 9001, shopName: 'Arms', items: [] });
             playerStub.getCurrentShop.returns(shop);
 
-            await service.closeShop(playerStub);
+            await manager.closeShop(playerStub);
 
             expect(playerStub.setCurrentShop.calledOnceWith(null)).to.be.true;
             expect(playerStub.sendShopClose.calledOnce).to.be.true;
@@ -91,7 +91,7 @@ describe('ShopService', () => {
         it('should do nothing if player has no open shop', async () => {
             playerStub.getCurrentShop.returns(null);
 
-            await service.closeShop(playerStub);
+            await manager.closeShop(playerStub);
 
             expect(playerStub.setCurrentShop.called).to.be.false;
             expect(playerStub.sendShopClose.called).to.be.false;
@@ -102,7 +102,7 @@ describe('ShopService', () => {
         it('should not buy if player has no open shop', async () => {
             playerStub.getCurrentShop.returns(null);
 
-            await service.buy(playerStub, 0);
+            await manager.buy(playerStub, 0);
 
             expect(playerStub.addPoint.called).to.be.false;
             expect(playerStub.sendShopResult.called).to.be.false;
@@ -112,7 +112,7 @@ describe('ShopService', () => {
             const shop = new Shop({ npcVnum: 9001, shopName: 'Arms', items: [] });
             playerStub.getCurrentShop.returns(shop);
 
-            await service.buy(playerStub, 5);
+            await manager.buy(playerStub, 5);
 
             expect(playerStub.sendShopResult.calledOnceWith({ result: ShopSubHeaderGC.INVALID_POS })).to.be.true;
         });
@@ -125,7 +125,7 @@ describe('ShopService', () => {
             playerStub.getCurrentShop.returns(shop);
             playerStub.getPoint.withArgs(PointsEnum.GOLD).returns(100);
 
-            await service.buy(playerStub, 0);
+            await manager.buy(playerStub, 0);
 
             expect(playerStub.sendShopResult.calledOnceWith({ result: ShopSubHeaderGC.NOT_ENOUGH_MONEY })).to.be.true;
             expect(playerStub.addPoint.called).to.be.false;
@@ -140,7 +140,7 @@ describe('ShopService', () => {
             playerStub.getPoint.withArgs(PointsEnum.GOLD).returns(1000);
             itemManagerStub.getItem.returns(null);
 
-            await service.buy(playerStub, 0);
+            await manager.buy(playerStub, 0);
 
             expect(playerStub.sendShopResult.calledOnceWith({ result: ShopSubHeaderGC.SOLD_OUT })).to.be.true;
         });
@@ -156,7 +156,7 @@ describe('ShopService', () => {
             itemManagerStub.getItem.returns(mockItem);
             playerStub.addItem.returns(false);
 
-            await service.buy(playerStub, 0);
+            await manager.buy(playerStub, 0);
 
             expect(playerStub.sendShopResult.calledOnceWith({ result: ShopSubHeaderGC.INVENTORY_FULL })).to.be.true;
             expect(playerStub.addPoint.called).to.be.false;
@@ -173,7 +173,7 @@ describe('ShopService', () => {
             itemManagerStub.getItem.returns(mockItem);
             playerStub.addItem.returns(true);
 
-            await service.buy(playerStub, 0);
+            await manager.buy(playerStub, 0);
 
             expect(playerStub.addPoint.calledOnceWith(PointsEnum.GOLD, -200)).to.be.true;
             expect(itemManagerStub.save.calledOnceWith(mockItem)).to.be.true;
@@ -185,7 +185,7 @@ describe('ShopService', () => {
         it('should not sell if player has no open shop', async () => {
             playerStub.getCurrentShop.returns(null);
 
-            await service.sell(playerStub, 1, 1);
+            await manager.sell(playerStub, 1, 1);
 
             expect(playerStub.addPoint.called).to.be.false;
         });
@@ -195,7 +195,7 @@ describe('ShopService', () => {
             playerStub.getCurrentShop.returns(shop);
             playerStub.getItem.returns(null);
 
-            await service.sell(playerStub, 1, 1);
+            await manager.sell(playerStub, 1, 1);
 
             expect(playerStub.sendShopResult.calledOnceWith({ result: ShopSubHeaderGC.INVALID_POS })).to.be.true;
         });
@@ -211,7 +211,7 @@ describe('ShopService', () => {
             playerStub.getCurrentShop.returns(shop);
             playerStub.getItem.returns(mockItem);
 
-            await service.sell(playerStub, 1, 1);
+            await manager.sell(playerStub, 1, 1);
 
             expect(playerStub.sendShopResult.calledOnceWith({ result: ShopSubHeaderGC.INVALID_POS })).to.be.true;
             expect(playerStub.addPoint.called).to.be.false;
@@ -230,7 +230,7 @@ describe('ShopService', () => {
             playerStub.getItem.returns(mockItem);
             playerStub.getInventory.returns(inventoryStub);
 
-            await service.sell(playerStub, 1, 3);
+            await manager.sell(playerStub, 1, 3);
 
             // sell price = floor(500 * 3 / 5) = 300
             expect(playerStub.addPoint.calledOnceWith(PointsEnum.GOLD, 300)).to.be.true;
@@ -249,7 +249,7 @@ describe('ShopService', () => {
             playerStub.getItem.returns(mockItem);
             playerStub.getInventory.returns(inventoryStub);
 
-            await service.sell(playerStub, 2, 1);
+            await manager.sell(playerStub, 2, 1);
 
             expect(inventoryStub.removeItem.calledOnceWith(2, 1)).to.be.true;
             expect(playerStub.sendItemRemoved.calledOnce).to.be.true;
@@ -270,7 +270,7 @@ describe('ShopService', () => {
             playerStub.getItem.returns(mockItem);
             playerStub.getInventory.returns(inventoryStub);
 
-            await service.sell(playerStub, 1, 10); // trying to sell 10, but only 2 available
+            await manager.sell(playerStub, 1, 10); // trying to sell 10, but only 2 available
 
             // sell price = floor(500 * 2 / 5) = 200
             expect(playerStub.addPoint.calledOnceWith(PointsEnum.GOLD, 200)).to.be.true;
